+++ src/shared/interface.ts
@@ -1,5 +1,4 @@
-// TODO: Remove import and properly declare global scope for interface Window
-import { QueryStringParams } from "../shared/types";
+export { }
 
 declare global {
   interface Window {
