+++ src/shared/types.ts
@@ -1,11 +1,12 @@
 
 export type QueryStringParams = {
-  appId?: string
-  mediajelAppId?: string
+  appId: string
+  mediajelAppId: string
   environment: string
   event: string
   test: string
-  version?: string
+  version: string
+  retailId: string
 }
