+++ src/shared/utils/try-parse-json.ts
@@ -1,0 +1,14 @@
+
+
+
+export const tryParseJSONObject = (str: string): string | any => {
+    try {
+        const o: any = JSON.parse(str);
+        if (o && typeof o === "object") {
+            return o;
+        }
+    }
+    catch (e) { }
+
+    return str;
+};
